added api's

+++ src/app/dashboard/components/todolist/todolist.component.ts
@@ -1,5 +1,5 @@
 import { Component, OnInit } from '@angular/core';
-import {CdkDragDrop, moveItemInArray, transferArrayItem} from '@angular/cdk/drag-drop';
+import { CdkDragDrop, moveItemInArray, transferArrayItem } from '@angular/cdk/drag-drop';
 import { MatDialog } from '@angular/material/dialog';
 import { AddpopupComponent } from '../addpopup/addpopup.component';
 import * as _ from 'underscore';
@@ -13,33 +13,41 @@ import { ActivatedRoute, Router } from '@angular/router';
   styleUrls: ['./todolist.component.scss']
 })
 export class TodolistComponent implements OnInit {
-  todo: listDto[]= [];
-  todayDate=new Date().toISOString();
-   userDetails: any = localStorage.getItem('userDetails') == null ? '' : localStorage.getItem('userDetails');
-   userId = JSON.parse(this.userDetails)['user']['_id'];
-   inprogress: listDto[]= [];
-   done: listDto[]= [];
-  constructor(public dialog: MatDialog,private dashboardService:DashboardService,
-    private router:Router,private activeRoute:ActivatedRoute) {
-    
-   }
+  todo: listDto[] = [];
+  todayDate = new Date().toISOString();
+  userDetails: any = localStorage.getItem('userDetails') == null ? '' : localStorage.getItem('userDetails');
+  userId = JSON.parse(this.userDetails)['user']['_id'];
+  inprogress: listDto[] = [];
+  done: listDto[] = [];
+  constructor(public dialog: MatDialog, private dashboardService: DashboardService,
+    private router: Router, private activeRoute: ActivatedRoute) {
+
+  }
 
   ngOnInit(): void {
-      this.getTodods();
+    this.getTodods();
   }
-  getTodods(){
-   this.activeRoute.params.subscribe(value=>{
-    this.todayDate = value['listDate'];
+  getTodods() {
+    this.activeRoute.params.subscribe(value => {
+      this.todayDate = value['listDate'];
     });
-    this.dashboardService.getTodoList(this.userId,new Date(this.todayDate).toISOString())
-    .subscribe(listTodo=>{
-      this.todo = listTodo.filter(el=>el.listStatus==1);
-      this.inprogress = listTodo.filter(el=>el.listStatus==2);
-      this.done = listTodo.filter(el=>el.listStatus==3);
-    });
+    this.dashboardService.getTodoList(this.userId, new Date(this.todayDate).toISOString())
+      .subscribe(listTodo => {
+        this.todo = listTodo.filter(el => el.listStatus == 1);
+        this.inprogress = listTodo.filter(el => el.listStatus == 2);
+        this.done = listTodo.filter(el => el.listStatus == 3);
+      });
   }
   droptodo(event: CdkDragDrop<any[]>) {
-    console.log(event.previousContainer.data[event.previousIndex]) 
+    if(event.previousContainer.data[event.previousIndex]){
+      event.previousContainer.data[event.previousIndex]['listStatus'] = 1;
+      //add update api
+      this.dashboardService.update_todo(this.userId, event.previousContainer.data[event.previousIndex]).subscribe(result=>{
+        if(result)
+        this.getTodods();
+      })
+    }
+    console.log(event.previousContainer.data[event.previousIndex])
     if (event.previousContainer === event.container) {
       moveItemInArray(event.container.data, event.previousIndex, event.currentIndex);
     } else {
@@ -51,10 +59,19 @@ export class TodolistComponent implements OnInit {
       );
     }
   }
-  dropInprogress(event: CdkDragDrop<any[]>){
+  dropInprogress(event: CdkDragDrop<any[]>) {
+    if(event.previousContainer.data[event.previousIndex]){
+      event.previousContainer.data[event.previousIndex]['listStatus'] = 2;
+      //add update api
+      this.dashboardService.update_todo(this.userId, event.previousContainer.data[event.previousIndex]).subscribe(result=>{
+        if(result)
+        this.getTodods();
+      })
+    }
     console.log(event.previousContainer.data[event.previousIndex]) // get which got changed
     if (event.previousContainer === event.container) {
       moveItemInArray(event.container.data, event.previousIndex, event.currentIndex);
+
     } else {
       transferArrayItem(
         event.previousContainer.data,
@@ -64,8 +81,16 @@ export class TodolistComponent implements OnInit {
       );
     }
   }
-  dropDone(event: CdkDragDrop<any[]>){
-    console.log(event.previousContainer.data[event.previousIndex]) 
+  dropDone(event: CdkDragDrop<any[]>) {
+    if (event.previousContainer.data[event.previousIndex]) {
+      event.previousContainer.data[event.previousIndex]['listStatus'] = 3;
+     //add update api
+     this.dashboardService.update_todo(this.userId, event.previousContainer.data[event.previousIndex]).subscribe(result=>{
+      if(result)
+      this.getTodods();
+    })
+    }
+    console.log(event.previousContainer.data[event.previousIndex])
     if (event.previousContainer === event.container) {
       moveItemInArray(event.container.data, event.previousIndex, event.currentIndex);
     } else {
@@ -76,40 +101,51 @@ export class TodolistComponent implements OnInit {
         event.currentIndex,
       );
     }
+
   }
-  AddTodo(){
-    const dialogRef = this.dialog.open(AddpopupComponent,{height: '550px',
-    width: '600px'});
-    dialogRef.afterClosed().subscribe(result=>{
-      if(result)
-      this.dashboardService.addTodo(result,this.userId).subscribe(addTodoResult=>{
-       
-      })
+  AddTodo() {
+    const dialogRef = this.dialog.open(AddpopupComponent, {
+      height: '550px',
+      width: '600px'
+    });
+    dialogRef.afterClosed().subscribe(result => {
+      let listDate = new Date(result.listDate).toISOString();
+      if (result)
+        this.dashboardService.addTodo(result, this.userId).subscribe(addTodoResult => {
+          if (addTodoResult) {
+            this.getTodods();
+          }
+        })
     })
   }
-  editTodo(item:any){
-    const dialogRef = this.dialog.open(AddpopupComponent,{
-      data:item,
+  editTodo(item: any) {
+    const dialogRef = this.dialog.open(AddpopupComponent, {
+      data: item,
       height: '550px',
-    width: '600px'});
-    dialogRef.afterClosed().subscribe(result=>{
-      if(result){
-        let index = this.todo.findIndex(data=>data['_id']==result['_id']);
-        this.todo.splice(index,1);
-        this.todo.push(result);
+      width: '600px'
+    });
+    dialogRef.afterClosed().subscribe(result => {
+      if (result) {
+        this.dashboardService.update_todo(this.userId, result).subscribe(result=>{
+          if(result)
+          this.getTodods();
+        })
       }
     })
   }
-  deleteTodo(item:any){
-    // const dialogRef = this.dialog.open(ConfirmationDialogComponent,{
-    //   data:item,
-    //   height:'200px',
-    //   width:'400px'
-    // })
-
-    this.dashboardService.delete_todo(this.userId,item._id).subscribe(result=>{
-      console.log(result);
-      this.getTodods()
+  deleteTodo(item: any) {
+    const dialogRef = this.dialog.open(ConfirmationDialogComponent,{
+      data:item,
+      height:'200px',
+      width:'400px'
+    }).afterClosed().subscribe(confirm=>{
+      if(confirm){
+        this.dashboardService.delete_todo(this.userId, item._id).subscribe(result => {
+          console.log(result);
+          this.getTodods()
+        })
+      }
     })
+
   }
 }
